Move `EdgeModal` to modals folder
