+++ src/components/scenes/existingAccout/SecurityAlertsScene.tsx
@@ -13,8 +13,11 @@ import { useImports } from '../../../hooks/useImports'
 import { useDispatch } from '../../../types/ReduxTypes'
 import { SceneProps } from '../../../types/routerTypes'
 import { toLocalTime } from '../../../util/utils'
+import { ButtonsView, StyledButtonContainer } from '../../buttons/ButtonsView'
+import { EdgeButton } from '../../buttons/EdgeButton'
 import { EdgeTouchableWithoutFeedback } from '../../common/EdgeTouchableWithoutFeedback'
 import { Space } from '../../layout/Space'
+import { EdgeModal } from '../../modals/EdgeModal'
 import { Airship, showError } from '../../services/AirshipInstance'
 import { Theme, ThemeProps, useTheme } from '../../services/ThemeContext'
 import { EdgeText, Paragraph, WarningText } from '../../themed/EdgeText'
@@ -24,10 +27,7 @@ import { ModalTitle } from '../../themed/ModalParts'
 import { ThemedScene } from '../../themed/ThemedScene'
 import { MessageText } from '../../themed/ThemedText'
 import { AlertCard } from '../../ui4/AlertCard'
-import { ButtonsView, StyledButtonContainer } from '../../buttons/ButtonsView'
-import { EdgeButton } from '../../buttons/EdgeButton'
 import { EdgeCard } from '../../ui4/EdgeCard'
-import { EdgeModal } from '../../ui4/EdgeModal'
 
 export interface SecurityAlertParams {
   account: EdgeAccount
